# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Monitorban API",
        "description": "Monitorban REST API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current host"
        }
    ],
    "paths": {
...
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user account",
                "operationId": "cff73fd071f3310eaa857e8674331eaf",
                "requestBody": {
                    "description": "User registration payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registration successful"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "login user account",
                "operationId": "2d4442d06a644df0fb22e8de8a5a93dd",
                "requestBody": {
                    "description": "User Login payload",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequestDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login successful"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/LoginRequestDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/verify-otp": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verify the one-time password (OTP)",
                "operationId": "a061564c8d1b9544442483f7d966f510",
                "requestBody": {
                    "description": "Verify OTP for authentication",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile",
                                    "otp"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "string",
                                        "example": "09123456789"
                                    },
                                    "otp": {
                                        "type": "integer",
                                        "example": 123456
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP verified successfully"
                    },
                    "400": {
                        "description": "Invalid or expired OTP"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/v1/auth/request-otp": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Request a one-time password (OTP)",
                "operationId": "bb1076cdaa8505ee7200416db1453eef",
                "requestBody": {
                    "description": "Request OTP for authentication",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "mobile"
                                ],
                                "properties": {
                                    "mobile": {
                                        "type": "string",
                                        "example": "09371801145"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP generated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/v1/auth/refresh-token": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Refresh the access token using a refresh token",
                "operationId": "f7baa8d2c3baf5fa4cbd70e0e833347b",
                "requestBody": {
                    "description": "Refresh token details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "refresh_token"
                                ],
                                "properties": {
                                    "refresh_token": {
                                        "description": "The refresh token used to obtain a new access token.",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "New access token generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "expires_in": {
                                            "description": "Time in seconds until the token expires",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or expired refresh token"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LoginRequestDTO": {
                "properties": {
                    "mobile": {
                        "type": "string",
                        "example": "09123456789"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    }
                },
                "type": "object"
            },
...
            "UserRequestDTO": {
                "properties": {
                    "first_name": {
                        "type": "string",
                        "example": "محمدرضا"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "گوهری"
                    },
                    "username": {
                        "type": "string",
                        "example": "gohari"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "example@gandom.link"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "password123"
                    },
                    "mobile": {
                        "type": "string",
                        "example": "09123456789"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token as: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}